move ChatAnthropicBedrock to its own module
