+++ libs/partners/anthropic/langchain_anthropic/__init__.py
@@ -1,9 +1,9 @@
 """Claude (Anthropic) partner package for LangChain."""
 
 from langchain_anthropic._version import __version__
+from langchain_anthropic.bedrock import ChatAnthropicBedrock
 from langchain_anthropic.chat_models import (
     ChatAnthropic,
-    ChatAnthropicBedrock,
     convert_to_anthropic_tool,
 )
 from langchain_anthropic.llms import AnthropicLLM
